Extract user overlay actions button component (#36236)
This PR extracts the classes used for the actions dropdown on the user
profile overlay to separate components, in preparation to be used on the
team overlay (#32584).

Kinda RFC since I'm not sure if this is the best way to do this.
